pyproject.toml
[tool.poetry]
name = "git-gh-automation"
version = "0.1.0"
description = "Git automation of GitHub"
authors = ["Przemyslaw Lukasik <przemyslaw.lukasik@vidaa.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.4"
playwright = "^1.50.0"
pytest-playwright = "^0.7.0"
black = "^25.1.0"
flake8 = "^7.1.1"
isort = "^6.0.0"
python-dotenv = "^1.0.1"
pytest-dotenv = "^0.5.2"
gitpython = "^3.1.44"
pytest-xdist = "^3.6.1"
pytest-logger = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''

(
  /(
     \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | docs
  )/

)
'''

[tool.isort]
profile = "black"
multi_line_output = 3